Move Statuspage.io config to a dedicated file
